+++ core/src/deserialize_key.rs
@@ -12,27 +12,13 @@ pub trait Sealed: Sized {
     fn skip<B: Buf>(buf: &mut B) -> Result<(), DeserializeError>;
 }
 
-pub trait DeserializeKey: Sealed + Sized {}
+pub trait DeserializeKeyImpl: Sealed {}
 
-impl<T: DeserializeKey> Sealed for Box<T> {
-    fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
-        T::deserialize_key(buf).map(Box::new)
-    }
+pub trait DeserializeKey: Sized {
+    type Impl: DeserializeKeyImpl;
 
-    fn deserialize_map_value_kind<B: Buf>(buf: &mut B) -> Result<(), DeserializeError> {
-        T::deserialize_map_value_kind(buf)
-    }
-
-    fn deserialize_set_value_kind<B: Buf>(buf: &mut B) -> Result<(), DeserializeError> {
-        T::deserialize_set_value_kind(buf)
-    }
-
-    fn skip<B: Buf>(buf: &mut B) -> Result<(), DeserializeError> {
-        T::skip(buf)
-    }
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError>;
 }
-
-impl<T: DeserializeKey> DeserializeKey for Box<T> {}
 
 impl Sealed for u8 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -52,7 +38,15 @@ impl Sealed for u8 {
     }
 }
 
-impl DeserializeKey for u8 {}
+impl DeserializeKeyImpl for u8 {}
+
+impl DeserializeKey for u8 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for i8 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -72,7 +66,15 @@ impl Sealed for i8 {
     }
 }
 
-impl DeserializeKey for i8 {}
+impl DeserializeKeyImpl for i8 {}
+
+impl DeserializeKey for i8 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for u16 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -92,7 +94,15 @@ impl Sealed for u16 {
     }
 }
 
-impl DeserializeKey for u16 {}
+impl DeserializeKeyImpl for u16 {}
+
+impl DeserializeKey for u16 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for i16 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -112,7 +122,15 @@ impl Sealed for i16 {
     }
 }
 
-impl DeserializeKey for i16 {}
+impl DeserializeKeyImpl for i16 {}
+
+impl DeserializeKey for i16 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for u32 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -132,7 +150,15 @@ impl Sealed for u32 {
     }
 }
 
-impl DeserializeKey for u32 {}
+impl DeserializeKeyImpl for u32 {}
+
+impl DeserializeKey for u32 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for i32 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -152,7 +178,15 @@ impl Sealed for i32 {
     }
 }
 
-impl DeserializeKey for i32 {}
+impl DeserializeKeyImpl for i32 {}
+
+impl DeserializeKey for i32 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for u64 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -172,7 +206,15 @@ impl Sealed for u64 {
     }
 }
 
-impl DeserializeKey for u64 {}
+impl DeserializeKeyImpl for u64 {}
+
+impl DeserializeKey for u64 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for i64 {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -192,7 +234,15 @@ impl Sealed for i64 {
     }
 }
 
-impl DeserializeKey for i64 {}
+impl DeserializeKeyImpl for i64 {}
+
+impl DeserializeKey for i64 {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for String {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -215,7 +265,15 @@ impl Sealed for String {
     }
 }
 
-impl DeserializeKey for String {}
+impl DeserializeKeyImpl for String {}
+
+impl DeserializeKey for String {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
 
 impl Sealed for Uuid {
     fn deserialize_key<B: Buf>(buf: &mut B) -> Result<Self, DeserializeError> {
@@ -237,4 +295,20 @@ impl Sealed for Uuid {
     }
 }
 
-impl DeserializeKey for Uuid {}
+impl DeserializeKeyImpl for Uuid {}
+
+impl DeserializeKey for Uuid {
+    type Impl = Self;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        Ok(key)
+    }
+}
+
+impl<T: DeserializeKey> DeserializeKey for Box<T> {
+    type Impl = T::Impl;
+
+    fn try_from_impl(key: Self::Impl) -> Result<Self, DeserializeError> {
+        T::try_from_impl(key).map(Self::new)
+    }
+}