+++ core/src/serialize_key.rs
@@ -4,65 +4,25 @@ use crate::value::ValueKind;
 use bytes::BufMut;
 use uuid::Uuid;
 
-pub trait Sealed {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError>;
+pub trait Sealed: Sized {
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError>;
     fn serialize_map_value_kind<B: BufMut>(buf: &mut B);
     fn serialize_set_value_kind<B: BufMut>(buf: &mut B);
 }
 
-pub trait SerializeKey: Sealed {}
+pub trait SerializeKeyImpl: Sealed {}
 
-impl<T: Sealed + ?Sized> Sealed for &T {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        (**self).serialize_key(buf)
-    }
+pub trait SerializeKey {
+    type Impl<'a>: SerializeKeyImpl
+    where
+        Self: 'a;
 
-    fn serialize_map_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_map_value_kind(buf)
-    }
-
-    fn serialize_set_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_set_value_kind(buf)
-    }
+    fn as_impl(&self) -> Self::Impl<'_>;
 }
-
-impl<T: SerializeKey + ?Sized> SerializeKey for &T {}
-
-impl<T: Sealed + ?Sized> Sealed for &mut T {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        (**self).serialize_key(buf)
-    }
-
-    fn serialize_map_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_map_value_kind(buf)
-    }
-
-    fn serialize_set_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_set_value_kind(buf)
-    }
-}
-
-impl<T: SerializeKey + ?Sized> SerializeKey for &mut T {}
-
-impl<T: Sealed + ?Sized> Sealed for Box<T> {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        (**self).serialize_key(buf)
-    }
-
-    fn serialize_map_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_map_value_kind(buf)
-    }
-
-    fn serialize_set_value_kind<B: BufMut>(buf: &mut B) {
-        T::serialize_set_value_kind(buf)
-    }
-}
-
-impl<T: SerializeKey + ?Sized> SerializeKey for Box<T> {}
 
 impl Sealed for u8 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_u8(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_u8(self);
         Ok(())
     }
 
@@ -75,11 +35,19 @@ impl Sealed for u8 {
     }
 }
 
-impl SerializeKey for u8 {}
+impl SerializeKeyImpl for u8 {}
+
+impl SerializeKey for u8 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for i8 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_i8(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_i8(self);
         Ok(())
     }
 
@@ -92,11 +60,19 @@ impl Sealed for i8 {
     }
 }
 
-impl SerializeKey for i8 {}
+impl SerializeKeyImpl for i8 {}
+
+impl SerializeKey for i8 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for u16 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_u16_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_u16_le(self);
         Ok(())
     }
 
@@ -109,11 +85,19 @@ impl Sealed for u16 {
     }
 }
 
-impl SerializeKey for u16 {}
+impl SerializeKeyImpl for u16 {}
+
+impl SerializeKey for u16 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for i16 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_i16_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_i16_le(self);
         Ok(())
     }
 
@@ -126,11 +110,19 @@ impl Sealed for i16 {
     }
 }
 
-impl SerializeKey for i16 {}
+impl SerializeKeyImpl for i16 {}
+
+impl SerializeKey for i16 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for u32 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_u32_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_u32_le(self);
         Ok(())
     }
 
@@ -143,11 +135,19 @@ impl Sealed for u32 {
     }
 }
 
-impl SerializeKey for u32 {}
+impl SerializeKeyImpl for u32 {}
+
+impl SerializeKey for u32 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for i32 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_i32_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_i32_le(self);
         Ok(())
     }
 
@@ -160,11 +160,19 @@ impl Sealed for i32 {
     }
 }
 
-impl SerializeKey for i32 {}
+impl SerializeKeyImpl for i32 {}
+
+impl SerializeKey for i32 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for u64 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_u64_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_u64_le(self);
         Ok(())
     }
 
@@ -177,11 +185,19 @@ impl Sealed for u64 {
     }
 }
 
-impl SerializeKey for u64 {}
+impl SerializeKeyImpl for u64 {}
+
+impl SerializeKey for u64 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
 
 impl Sealed for i64 {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        buf.put_varint_i64_le(*self);
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
+        buf.put_varint_i64_le(self);
         Ok(())
     }
 
@@ -194,10 +210,18 @@ impl Sealed for i64 {
     }
 }
 
-impl SerializeKey for i64 {}
+impl SerializeKeyImpl for i64 {}
 
-impl Sealed for str {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
+impl SerializeKey for i64 {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
+
+impl Sealed for &str {
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
         if self.len() <= u32::MAX as usize {
             buf.put_varint_u32_le(self.len() as u32);
             buf.put_slice(self.as_bytes());
@@ -216,26 +240,21 @@ impl Sealed for str {
     }
 }
 
-impl SerializeKey for str {}
+impl SerializeKeyImpl for &str {}
 
-impl Sealed for String {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
-        self.as_str().serialize_key(buf)
-    }
+impl SerializeKey for &str {
+    type Impl<'a>
+        = Self
+    where
+        Self: 'a;
 
-    fn serialize_map_value_kind<B: BufMut>(buf: &mut B) {
-        str::serialize_map_value_kind(buf);
-    }
-
-    fn serialize_set_value_kind<B: BufMut>(buf: &mut B) {
-        str::serialize_set_value_kind(buf);
+    fn as_impl(&self) -> Self::Impl<'_> {
+        self
     }
 }
 
-impl SerializeKey for String {}
-
 impl Sealed for Uuid {
-    fn serialize_key<B: BufMut>(&self, buf: &mut B) -> Result<(), SerializeError> {
+    fn serialize_key<B: BufMut>(self, buf: &mut B) -> Result<(), SerializeError> {
         buf.put_slice(self.as_bytes());
         Ok(())
     }
@@ -249,4 +268,53 @@ impl Sealed for Uuid {
     }
 }
 
-impl SerializeKey for Uuid {}
+impl SerializeKeyImpl for Uuid {}
+
+impl SerializeKey for Uuid {
+    type Impl<'a> = Self;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        *self
+    }
+}
+
+impl<T: SerializeKey + ?Sized> SerializeKey for &T {
+    type Impl<'a>
+        = T::Impl<'a>
+    where
+        Self: 'a;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        (**self).as_impl()
+    }
+}
+
+impl<T: SerializeKey + ?Sized> SerializeKey for &mut T {
+    type Impl<'a>
+        = T::Impl<'a>
+    where
+        Self: 'a;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        (**self).as_impl()
+    }
+}
+
+impl<T: SerializeKey + ?Sized> SerializeKey for Box<T> {
+    type Impl<'a>
+        = T::Impl<'a>
+    where
+        Self: 'a;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        (**self).as_impl()
+    }
+}
+
+impl SerializeKey for String {
+    type Impl<'a> = &'a str;
+
+    fn as_impl(&self) -> Self::Impl<'_> {
+        self
+    }
+}